Code & docs review for server instrumentation patch.  File timestamps should surely be timestamptz not timestamp; fix some but not all of the holes in check_and_make_absolute(); other minor cleanup.  Also put in the missed catversion bump.

+++ src/backend/postmaster/postmaster.c
@@ -37,7 +37,7 @@
  *
  *
  * IDENTIFICATION
- *	  $PostgreSQL: pgsql/src/backend/postmaster/postmaster.c,v 1.463 2005/08/12 03:23:51 momjian Exp $
+ *	  $PostgreSQL: pgsql/src/backend/postmaster/postmaster.c,v 1.464 2005/08/12 18:23:53 tgl Exp $
  *
  * NOTES
  *
@@ -3382,20 +3382,22 @@ sigusr1_handler(SIGNAL_ARGS)
 		}
 	}
 
-	if (PgArchPID != 0 && Shutdown == NoShutdown)
+	if (CheckPostmasterSignal(PMSIGNAL_WAKEN_ARCHIVER) &&
+		PgArchPID != 0 && Shutdown == NoShutdown)
 	{
-		if (CheckPostmasterSignal(PMSIGNAL_WAKEN_ARCHIVER))
-		{
-			/*
-			 * Send SIGUSR1 to archiver process, to wake it up and begin
-			 * archiving next transaction log file.
-			 */
-			kill(PgArchPID, SIGUSR1);
-		}
+		/*
+		 * Send SIGUSR1 to archiver process, to wake it up and begin
+		 * archiving next transaction log file.
+		 */
+		kill(PgArchPID, SIGUSR1);
 	}
 
-	if (CheckPostmasterSignal(PMSIGNAL_ROTATE_LOGFILE) && SysLoggerPID != 0)
+	if (CheckPostmasterSignal(PMSIGNAL_ROTATE_LOGFILE) &&
+		SysLoggerPID != 0)
+	{
+		/* Tell syslogger to rotate logfile */
 	    kill(SysLoggerPID, SIGUSR1);
+	}
 
 	PG_SETMASK(&UnBlockSig);
